requirements.txt
torch
torchvision
requests
numpy
pandas
python-box
PyYAML